segundo commit

+++ mern-proyect/src/components/books/booksList.jsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from 'react';
-import { getBooks, deleteBook } from './services/api.js';
+import { getBooks, deleteBook } from '../../services/api.js';
 
 const BooksList = () => {
     const [books, setBooks] = useState([]);
@@ -17,30 +17,33 @@ const BooksList = () => {
     }, []);
 
     const handleDelete = async (id) => {
-        try {
-            await deleteBook(id);
-            setBooks(books.filter((book) => book.id !== id));
-        } catch (error) {
-            console.error('Error al eliminar libro:', error);
-        }
+        const confirm=window.confirm("¿Estás seguro de querer borrar este libro?")
+        if (confirm){
+            try {
+                await deleteBook(id);
+                setBooks(books.filter((book) => book.id !== id));
+            } catch (error) {
+                console.error('Error al eliminar libro:', error);
+                alert("Error al eliminar libro");//notificiación al usuario
+            }
+        } 
     };
 
     return (
         <div>
-            <h2>Lista de Libros</h2>
-            <ul>
-                {books.map((book)=> (
-                    <li key={book.id}>
-                        <h3>{book.title}</h3>
-                        <h3>{book.author}</h3>
-                        <p>{book.genre}</p>
-                        <small>Publicado el: {new Date(book.publishedAt).toLocaleDateString()}</small>
-                        <button onClick={() => handleDelete(book.id)}>Eliminar</button>
-                    </li>
-                ))}
-            </ul>
+            <h1>Lista de Libros</h1>
+            {books.map((book) => (
+                <div key={book._id}>
+                    <h2>{book.title}</h2>
+                    <h3>{book.author}</h3>
+                    <h3>{book.genre}</h3>
+                    <small>Publicación: {new Date(book.publishedAt).toLocaleDateString()}</small>
+                    <button onClick={() => handleDelete(book._id)}>Eliminar Libro</button>
+                </div>
+            ))}
         </div>
     );
+    
 };
 
 export default BooksList;
